sales-backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-multipart
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
python-dotenv
psycopg2-binary
google-generativeai
